.walkingkooka-j2cl-maven-plugin-ignored-files.txt was .walkingkooka-j2cl-maven-plugin-ignore.txt
- only filename changed, no logic changes.
