+++ src/it/example-ignored-files/pom.xml
@@ -3,7 +3,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>walkingkooka</groupId>
-    <artifactId>example-ignore-file</artifactId>
+    <artifactId>example-ignored-files</artifactId>
     <version>1.0</version>
     <packaging>war</packaging>
 
@@ -74,7 +74,7 @@
                                 <jre.checks.checkLevel>MINIMAL</jre.checks.checkLevel>
                                 <jsinterop.checks>DISABLED</jsinterop.checks>
                             </defines>
-                            <entry-points>example.ignorefile.app</entry-points>
+                            <entry-points>example.ignoredfiles.app</entry-points>
                             <externs></externs>
                             <formatting/>
                             <language-out>ECMASCRIPT_2016</language-out>
