+++ src/it/example-ignored-files/src/main/java/example/ignoredfileslib/.walkingkooka-j2cl-maven-plugin-ignored-files.txt
