+++ src/it/example-ignored-files/src/main/webapp/WEB-INF/web.xml
