+++ src/it/example-ignored-files/src/main/webapp/index.html
