+++ src/main/java/walkingkooka/j2cl/maven/J2clPath.java
@@ -258,14 +258,14 @@ final class J2clPath implements Comparable<J2clPath> {
     /**
      * Builds a new path holding the ignore file.
      */
-    J2clPath ignoreFile() {
-        return this.append(IGNORE_FILE);
+    J2clPath ignoredFiles() {
+        return this.append(IGNORED_FILES);
     }
 
     /**
      * The name of the ignore file which is used during the unpack phase to filter files.
      */
-    private static final String IGNORE_FILE = FILE_PREFIX + "-ignore.txt";
+    private static final String IGNORED_FILES = FILE_PREFIX + "-ignored-files.txt";
 
     /**
      * Returns true if this file is a java file.
